requirements.txt
django==1.3.1
django-mptt==0.5.1
south>=0.7.2
PIL==1.1.7
django-cms>=2.2
django-reversion==1.4
django-filer
cmsplugin-filer
django-loginza
flup
django-tagging>=0.3.1
BeautifulSoup>=3.2.0
django-xmlrpc>=0.1.3
pyparsing>=1.5.5
django-blog-zinnia
django-tinymce
django-simple-captcha
